Populate form values and move submission to a separate controller
