+++ composer.json
@@ -18,7 +18,7 @@
     ],
     "require": {
         "silverstripe/recipe-cms": "^4",
-        "silverstripe/userforms": "^5.1.1",
+        "silverstripe/userforms": "^5.4",
         "symbiote/silverstripe-queuedjobs": "^4.2"
     },
     "require-dev": {
