+++ templates/Firesphere/PartialUserforms/Layout/PartialUserForm.ss
@@ -1,14 +1,11 @@
 <div class="container">
-    <% if $Breadcrumbs %>
-        <% include Breadcrumbs Breadcrumbs=$Breadcrumbs %>
-    <% end_if %>
     <h1>$Title</h1>
 
-    <% if $Link %>
+    <% if $PartialLink %>
         <p>
             This form allows for multiple sessions (singular or multiple users).
             You can access or contribute to this form using the following link:
-            <a href="$Link">$Link</a>
+            <a href="$PartialLink">$PartialLink</a>
         </p>
     <% end_if %>
