+++ tests/unit/PartialFieldSubmissionTest.php
@@ -7,6 +7,7 @@ use Firesphere\PartialUserforms\Models\PartialFormSubmission;
 use SilverStripe\Dev\SapphireTest;
 use SilverStripe\Security\DefaultAdminService;
 use SilverStripe\Security\Member;
+use SilverStripe\Security\Security;
 use SilverStripe\UserForms\Model\UserDefinedForm;
 
 class PartialFieldSubmissionTest extends SapphireTest
@@ -26,46 +27,48 @@ class PartialFieldSubmissionTest extends SapphireTest
         $this->assertTrue($this->field->canView());
 
         $member = DefaultAdminService::singleton()->findOrCreateAdmin('admin@example.com');
-        $this->logInAs($member);
+        Security::setCurrentUser($member);
 
         $this->assertTrue($this->field->canView($member));
     }
 
     public function testCanCreate()
     {
-        Member::currentUser()->logOut();
+        Security::setCurrentUser(null);
         $this->assertFalse($this->field->canCreate());
 
         $member = DefaultAdminService::singleton()->findOrCreateAdmin('admin@example.com');
-        $this->logInAs($member);
+        Security::setCurrentUser($member);
 
-        $this->assertTrue($this->field->canCreate($member));
+        $this->assertFalse($this->field->canCreate($member));
     }
 
     public function testCanEdit()
     {
-        Member::currentUser()->logOut();
+        Security::setCurrentUser(null);
         $this->assertFalse($this->field->canEdit());
 
         $member = DefaultAdminService::singleton()->findOrCreateAdmin('admin@example.com');
-        $this->logInAs($member);
+        Security::setCurrentUser($member);
 
         $this->assertTrue($this->field->canEdit($member));
     }
 
     public function testCanDelete()
     {
-        Member::currentUser()->logOut();
+        Security::setCurrentUser(null);
         $this->assertFalse($this->field->canDelete());
 
         $member = DefaultAdminService::singleton()->findOrCreateAdmin('admin@example.com');
-        $this->logInAs($member);
+        Security::setCurrentUser($member);
 
         $this->assertTrue($this->field->canDelete($member));
     }
 
     protected function setUp()
     {
+        parent::setUp();
+
         $this->field = PartialFieldSubmission::create();
         $partialForm = PartialFormSubmission::create();
         $udf = UserDefinedForm::create(['Title' => 'Test'])->write();
@@ -76,7 +79,5 @@ class PartialFieldSubmissionTest extends SapphireTest
         // testCanCreate() will most likely be an error due to SubmittedForm returning a parent's permission
         // without checking if Parent exists, so have to set the SubmittedForm parent
         $this->field->SubmittedForm()->ParentID = $udf;
-
-        return parent::setUp();
     }
 }
